Move require_selinux to pulp_smash namespace.
Function `require_selinux` can be re-used in other tests later on.
Besides that, the function is not related to a specific Pulp
version.
